+++ packages/extensions/core/lib/configuration/configuration-loader.ts
